refactor(donation-widget): update stories to have isolated per each widget piece
all still live under the "Example/DonationWidget" so the UI still group them but code is easier to
use
